requirements.txt
# Dependências principais
python-binance==1.0.16
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1

# Machine Learning
scikit-learn==1.2.2
joblib==1.2.0

# Análise técnica
ta-lib==0.4.26

# Utilitários
tqdm==4.65.0
requests==2.29.0
python-dotenv==1.0.0